backend-rag/requirements.txt
# Backend RAG - Scriptorium AI
# Python 3.11+

# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# OpenAI (embeddings y LLM)
openai==1.10.0

# Google Cloud (Document AI, Vector Search, Storage)
google-cloud-documentai==2.24.0
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.14.0

# Vector DB alternativas (por si usamos Pinecone/Qdrant)
pinecone-client==3.0.0
qdrant-client==1.7.0

# Procesamiento de texto
tiktoken==0.5.2
langchain==0.1.0
langchain-openai==0.0.2
pypdf==4.0.0
python-docx==1.1.0
pytesseract==0.3.10

# Utilidades
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
